No scrollytelling yet; maps placed
I have re-configured the grid so that the right side shows a single map, while the left side scrolls. That means that each text section no longer calls its own map. In turn that m
eans the function for layer is no longer `mapify()` but `create_layer_function()`, because the function needs to be added to (piped from) an existing map. In  the process I also e
dited the `plot_basemap()` function to use CartoDB.Positron as the vector basemap. Likely to change, but at least it works – unlikely Wikimedia's.