Add --implicit-dirs to gcsfuse in gcsfuse_run.sh
May want to replace with a script that mkdirs all of the bucket paths. See link below. The commented out code in gcsfuse_run.sh is a failed attempt to do so.

This is necessary because of Cloud Storage's flat hierarchy: foo/bar/ does not mean bar/ is located within foo, it is simply named "foo/bar/". --implicit-dirs tells gcsfuse to treat foo/bar/ like it is a subdirectory of foo/. According to https://github.com/GoogleCloudPlatform/gcsfuse/blob/master/docs/semantics.md#files-and-directories there are some issues with this, so it may be worth looking into alternatives.